refactor(user-type): Moved User interface definition in server to its own file
