+++ server/models/User.ts
@@ -1,31 +1,14 @@
-import { Collection, FindCursor, MatchKeysAndValues, ObjectId, WithId } from 'mongodb';
+import { 
+  Collection, 
+  FindCursor, 
+  MatchKeysAndValues, 
+  ObjectId, 
+  WithId 
+} from 'mongodb';
 
 import { connectToDatabase } from '../config/db';
 
-import { Name } from '../types/interfaces/Name';
-import Permission from '../types/enums/Permission';
-import { UserAccountType } from '../types/types/UserAccountType';
-
-interface User {
-  _id?: string | null | ObjectId;
-  name?: Name;
-  password: string;
-  email: string;
-  phone?: string;
-  accType: UserAccountType;
-  // For non-Business User accounts, indicates the _id of 
-  // a Business User account they're associated with
-  businessId?: string;
-  businessName?: string;
-  businessWebsite?: string;
-  articlePermissions?: Permission;
-  userPermissions?: Permission;
-  hasDisabledLogin?: boolean;
-  // For Business Accounts, Indicates whether or not the account should receive
-  // an email about new comment notifcation
-  receivesCommentNotifications?: boolean;
-  notes?: string;
-}
+import { User } from '../types/interfaces/User';
 
 /**
  * @returns {Promise<Collection<User>>} A Promise meant to return the DB's
